refactor: Reorganize diet and database code into modules
- Move frontend diet code to src/diet/index.ts
- Create src-tauri/src/db/ module (DbPath, migrations)
- Create src-tauri/src/diet/ module (models, commands)
- Reduce lib.rs from 413 to 42 lines
- Add hierarchical CLAUDE.md files for each module
