+++ src-tauri/src/db/mod.rs
@@ -1,0 +1,4 @@
+pub mod migrations;
+
+/// Holds the database path for use in commands
+pub struct DbPath(pub String);
