+++ src-tauri/src/diet/mod.rs
@@ -1,0 +1,6 @@
+pub mod commands;
+pub mod models;
+
+// Re-export commonly used items
+pub use commands::*;
+pub use models::*;
